voice-agent/requirements.txt
livekit-agents>=1.4.5
livekit-plugins-anthropic>=1.3.0
livekit-plugins-deepgram>=1.3.0
livekit-plugins-elevenlabs>=1.3.0
livekit-plugins-silero>=1.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
